requirements.txt
Django>=6.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
django-ratelimit>=4.1.0
drf-spectacular>=0.27.0

